+++ src/CategoriesContainer.jsx
@@ -6,7 +6,7 @@ import MenuItem from './MenuItem';
 import {
   selectCategory,
   loadRestaurants,
-} from './actions';
+} from './slice';
 
 import { get } from './utils';
